# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrEnvelope:
    properties:
      error:
        type: string
    type: object
  handler.MsgEnvelope:
    properties:
      message:
        type: string
    type: object
  model.Book:
    properties:
      author:
        type: string
      id:
        minimum: 1
        type: integer
      title:
        type: string
      version:
        type: integer
    required:
    - id
    type: object
...
  model.CreateBookRq:
    properties:
      author:
        minLength: 5
        type: string
      title:
        minLength: 5
        type: string
    required:
    - author
    - title
    type: object
...
  model.UpdateBookRq:
    properties:
      author:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Onelab project
  title: Onelab API
  version: "1.0"
paths:
  /book:
    post:
      consumes:
      - application/json
      description: Create a new book with the input payload
      parameters:
      - description: Book information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateBookRq'
      produces:
      - application/json
      responses:
        "201":
          description: Book created
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "422":
          description: book already exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      summary: Create a new book
      tags:
      - book
  /book/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: book deleted
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: book is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Delete book by id
      tags:
      - book
    get:
      consumes:
      - application/json
      description: Get book by id in query param
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/model.Book'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: book is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Get book by id
      tags:
      - book
    patch:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book information
        in: body
        name: input
        schema:
          $ref: '#/definitions/model.UpdateBookRq'
      produces:
      - application/json
      responses:
        "200":
          description: book updated
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: book is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "409":
          description: edit conflict
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Update book by id
      tags:
      - book
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"